refactor: ApplyCard, GoToApply components move to components directory
