+++ src/components/styles/ApplyCard.css
