--- src/pages/IntroducePage/components/ApplyCardIntorduce.js
@@ -1,16 +1,0 @@
-import React from "react";
-import MainApplyCard from "../../MainPage/components/MainApplyCard";
-
-import "./styles/Introduce.css";
-import "./styles/ApplyCardIntroduce.css";
-
-const ApplyCardIntroduce = () => {
-  return (
-    <>
-      <div className="blank"></div>
-      <MainApplyCard />
-    </>
-  );
-};
-
-export default ApplyCardIntroduce;
