--- src/pages/IntroducePage/components/styles/ApplyCardIntroduce.css
@@ -1,3 +1,0 @@
-.blank{
-    margin-top: 235px;
-}
