+++ src/pages/IntroducePage/components/styles/ImgIntroduce.css
@@ -4,6 +4,8 @@
     justify-content: space-between;
     align-items: center;
     text-align: left;
+
+    height: 2079px;
 }
 
 .frame{
