+++ src/pages/MainPage/Main.js
@@ -3,14 +3,17 @@ import React from "react";
 // components
 import ImgMain from "./components/ImgMain/ImgMain";
 import IntroduceMain from "./components/IntroduceMain";
-import MainApplyCard from "./components/MainApplyCard";
+// import MainApplyCard from "./components/MainApplyCard";
+
+import ApplyCard from "../../components/ApplyCard";
 
 const Main = () => {
   return (
     <>
       <ImgMain />
       <IntroduceMain />
-      <MainApplyCard />
+      {/* <MainApplyCard /> */}
+      <ApplyCard />
     </>
   );
 };
